refactor(docs): radio-group dx
