+++ apps/docs/content/components/radio-group/custom-styles.raw.jsx
@@ -1,0 +1,36 @@
+import {RadioGroup, Radio, cn} from "@nextui-org/react";
+
+export const CustomRadio = (props) => {
+  const {children, ...otherProps} = props;
+
+  return (
+    <Radio
+      {...otherProps}
+      classNames={{
+        base: cn(
+          "inline-flex m-0 bg-content1 hover:bg-content2 items-center justify-between",
+          "flex-row-reverse max-w-[300px] cursor-pointer rounded-lg gap-4 p-4 border-2 border-transparent",
+          "data-[selected=true]:border-primary",
+        ),
+      }}
+    >
+      {children}
+    </Radio>
+  );
+};
+
+export default function App() {
+  return (
+    <RadioGroup description="Selected plan can be changed at any time." label="Plans">
+      <CustomRadio description="Up to 20 items" value="free">
+        Free
+      </CustomRadio>
+      <CustomRadio description="Unlimited items. $10 per month." value="pro">
+        Pro
+      </CustomRadio>
+      <CustomRadio description="24/7 support. Contact us for pricing." value="enterprise">
+        Enterprise
+      </CustomRadio>
+    </RadioGroup>
+  );
+}
