+++ apps/docs/content/components/radio-group/default-value.ts
@@ -1,20 +1,4 @@
-const App = `import {RadioGroup, Radio} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <RadioGroup
-      label="Select your favorite city"
-      color="secondary"
-      defaultValue="london"
-    >
-      <Radio value="buenos-aires">Buenos Aires</Radio>
-      <Radio value="sydney">Sydney</Radio>
-      <Radio value="san-francisco">San Francisco</Radio>
-      <Radio value="london">London</Radio>
-      <Radio value="tokyo">Tokyo</Radio>
-    </RadioGroup>
-  );
-}`;
+import App from "./default-value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
