+++ apps/docs/content/components/radio-group/horizontal.ts
@@ -1,19 +1,4 @@
-const App = `import {RadioGroup, Radio} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <RadioGroup
-      label="Select your favorite city"
-      orientation="horizontal"
-    >
-      <Radio value="buenos-aires">Buenos Aires</Radio>
-      <Radio value="sydney">Sydney</Radio>
-      <Radio value="san-francisco">San Francisco</Radio>
-      <Radio value="london">London</Radio>
-      <Radio value="tokyo">Tokyo</Radio>
-    </RadioGroup>
-  );
-}`;
+import App from "./horizontal.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
